+++ core/src/main/java/react4j/Context.java
@@ -113,7 +113,7 @@ public class Context<T>
   public static final class ConsumerBuilder<ST>
   {
     @Nonnull
-    private final ReactElement _element = ReactElement.createContextElement( React.Provider );
+    private final ReactElement _element = ReactElement.createContextElement( React.Consumer );
 
     private ConsumerBuilder()
     {